feat: change the path structure of the test files
