+++ src/components/cart_components/fordrawCart.ts
@@ -15,18 +15,3 @@ export function calculateDiscount(
 ): number {
   return Number((cartTotal * ((100 - discountCartTotal) / 100)).toFixed(2));
 }
-
-export function lengthWordsInput(arr: string[], a: number) {
-  return arr.every((item) => item.length >= a);
-}
-
-export function isIdProduct(arr: Product[], id: number) {
-  return arr.find((item) => item.id === id);
-}
-
-export function converteString(str: string) {
-  return str
-    .trim()
-    .split(" ")
-    .filter((item) => item !== "");
-}
