+++ src/components/product_components/forproduct.ts
@@ -1,0 +1,5 @@
+import { Product } from "../model/types";
+
+export function isIdProduct(arr: Product[], id: number) {
+  return arr.find((item) => item.id === id);
+}
